fix xo lint errors and convert ts->js
