+++ .eslintignore
@@ -1,0 +1,1 @@
+optimized-d3-cloud.js
