--- docs/ReactWordcloud.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import _ReactWordcloud from '../src';
-
-export default function ReactWordcloud(props) {
-  return <_ReactWordcloud {...props} />;
-}
-
-export * from '../src';
