+++ docs/common-issues.mdx
@@ -3,9 +3,9 @@ name: Common Issues
 route: /common-issues
 ---
 
-import { Playground } from 'docz';
+import { Playground } from "docz";
 
-import ReactWordcloud from './ReactWordcloud';
+import ReactWordcloud from "./react-wordcloud";
 
 # Common Issues
 
@@ -23,9 +23,9 @@ This issue happens when the most _frequent_ word is also the _longest_ word. For
   <ReactWordcloud
     size={[200, 200]}
     words={[
-      { text: 'this_is_a_long_word_and_also_the_most_frequent', value: 45 },
-      { text: 'some_word', value: 35 },
-      { text: 'another_word', value: 20 },
+      { text: "this_is_a_long_word_and_also_the_most_frequent", value: 45 },
+      { text: "some_word", value: 35 },
+      { text: "another_word", value: 20 },
     ]}
   />
 </Playground>
