--- docs/props/CallbacksProp.tsx
@@ -1,5 +1,0 @@
-import { CallbacksProp as Props } from '../../src/types';
-
-export default function CallbacksProp(_props: Props): JSX.Element {
-  return null;
-}
