+++ docs/props/props.mdx
@@ -3,11 +3,9 @@ name: Props
 route: /props
 ---
 
-import { Props } from 'docz';
+import { Props } from "docz";
 
-import CallbacksProp from './CallbacksProp';
-import OptionsProp from './OptionsProp';
-import ReactWordcloud from '../ReactWordcloud';
+import ReactWordcloud from "../react-wordcloud";
 
 # Props
 
@@ -17,10 +15,6 @@ import ReactWordcloud from '../ReactWordcloud';
 
 You can customize the wordcloud by configuring the `options` prop. All fields are optional.
 
-<Props of={OptionsProp} />
-
 ## `callbacks`
 
 The `callbacks` prop provides advanced ways to control the wordcloud behaviors. All fields are optional.
-
-<Props of={CallbacksProp} />
