+++ docs/react-wordcloud.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import ReactWordcloudSrc from '../src';
+
+const ReactWordcloud = props => {
+	return <ReactWordcloudSrc {...props} />;
+};
+
+export * from '../src';
+
+export default ReactWordcloud;
