+++ docs/usage/basic.mdx
@@ -3,10 +3,10 @@ route: /usage/basic
 menu: Usage
 ---
 
-import { Playground } from 'docz';
+import { Playground } from "docz";
 
-import ReactWordcloud from '../ReactWordcloud';
-import words from '../words';
+import ReactWordcloud from "../react-wordcloud";
+import words from "../words";
 
 # Basic Example
