+++ docs/usage/callbacks.mdx
@@ -4,11 +4,11 @@ route: /usage/callbacks
 menu: Usage
 ---
 
-import { select } from 'd3-selection';
-import { Playground } from 'docz';
+import { select } from "d3-selection";
+import { Playground } from "docz";
 
-import ReactWordcloud from '../ReactWordcloud';
-import words from '../words';
+import ReactWordcloud from "../react-wordcloud";
+import words from "../words";
 
 # Callbacks
 
@@ -25,7 +25,7 @@ In the example below, we can pass custom `getWordColor` and `getWordTooltip` cal
     words={words}
     callbacks={{
       getWordColor: ({ value }) => {
-        return value > 50 ? 'blue' : 'red';
+        return value > 50 ? "blue" : "red";
       },
       getWordTooltip: ({ text, value }) => {
         return value > 50 ? `${text.toUpperCase()} (> 50)` : `${text} (<= 50)`;
@@ -42,27 +42,27 @@ You can pass callbacks such as `onWordClick`, `onWordMouseEnter` and `onWordMous
   {() => {
     function getCallback(callbackName) {
       return function(word, event) {
-        const isActive = callbackName !== 'onWordMouseOut';
+        const isActive = callbackName !== "onWordMouseOut";
         const element = event.target;
         const text = select(element);
         text
-          .on('click', () => {
+          .on("click", () => {
             if (isActive) {
-              window.open(`https://duckduckgo.com/?q=${word.text}`, '_blank');
+              window.open(`https://duckduckgo.com/?q=${word.text}`, "_blank");
             }
           })
           .transition()
-          .attr('background', 'white')
-          .attr('font-size', isActive ? '300%' : '100%')
-          .attr('text-decoration', isActive ? 'underline' : 'none');
+          .attr("background", "white")
+          .attr("font-size", isActive ? "300%" : "100%")
+          .attr("text-decoration", isActive ? "underline" : "none");
       };
     }
     return (
       <ReactWordcloud
         callbacks={{
-          onWordClick: getCallback('onWordClick'),
-          onWordMouseOut: getCallback('onWordMouseOut'),
-          onWordMouseOver: getCallback('onWordMouseOver'),
+          onWordClick: getCallback("onWordClick"),
+          onWordMouseOut: getCallback("onWordMouseOut"),
+          onWordMouseOver: getCallback("onWordMouseOver"),
         }}
         words={words}
       />
