+++ docs/usage/optimizations.mdx
@@ -4,7 +4,7 @@ route: /usage/optimizations
 menu: Usage
 ---
 
-import ReactWordcloud from '../ReactWordcloud';
+import ReactWordcloud from '../react-wordcloud';
 import words from '../words';
 
 # Optimizations
