+++ docs/usage/options.mdx
@@ -4,10 +4,10 @@ route: /usage/options
 menu: Usage
 ---
 
-import { Playground } from 'docz';
+import { Playground } from "docz";
 
-import ReactWordcloud from '../ReactWordcloud';
-import words from '../words';
+import ReactWordcloud from "../react-wordcloud";
+import words from "../words";
 
 # Options
 
@@ -20,7 +20,7 @@ You can customize many visual and layout properties of `ReactWordcloud` by using
 <Playground>
   <ReactWordcloud
     options={{
-      colors: ['#1f77b4', '#9467bd', '#8c564b'],
+      colors: ["#1f77b4", "#9467bd", "#8c564b"],
     }}
     words={words}
   />
@@ -46,10 +46,10 @@ Configure wordfont styles using the `fontFamily`, `fontSizes`, `fontStyle`, `fon
 <Playground>
   <ReactWordcloud
     options={{
-      fontFamily: 'courier new',
+      fontFamily: "courier new",
       fontSizes: [10, 20],
-      fontStyle: 'italic',
-      fontWeight: 'bold',
+      fontStyle: "italic",
+      fontWeight: "bold",
     }}
     words={words}
   />
@@ -76,8 +76,8 @@ Configure the wordcloud layout by using the `scale`, `spiral` options.
 <Playground>
   <ReactWordcloud
     options={{
-      scale: 'log',
-      spiral: 'rectangular',
+      scale: "log",
+      spiral: "rectangular",
     }}
     words={words}
   />
@@ -92,24 +92,24 @@ Use the code editor to edit and play around with some of these options!
     words={words}
     options={{
       colors: [
-        '#1f77b4',
-        '#ff7f0e',
-        '#2ca02c',
-        '#d62728',
-        '#9467bd',
-        '#8c564b',
+        "#1f77b4",
+        "#ff7f0e",
+        "#2ca02c",
+        "#d62728",
+        "#9467bd",
+        "#8c564b",
       ],
       enableTooltip: true,
       deterministic: false,
-      fontFamily: 'impact',
+      fontFamily: "impact",
       fontSizes: [5, 60],
-      fontStyle: 'normal',
-      fontWeight: 'normal',
+      fontStyle: "normal",
+      fontWeight: "normal",
       padding: 1,
       rotations: 3,
       rotationAngles: [0, 90],
-      scale: 'sqrt',
-      spiral: 'archimedean',
+      scale: "sqrt",
+      spiral: "archimedean",
       transitionDuration: 1000,
     }}
   />
