+++ docs/usage/size.mdx
@@ -6,7 +6,7 @@ menu: Usage
 
 import { Playground } from 'docz';
 
-import ReactWordcloud from '../ReactWordcloud';
+import ReactWordcloud from '../react-wordcloud';
 import words from '../words';
 
 # Size
