+++ docs/usage/transitions.mdx
@@ -7,7 +7,7 @@ menu: Usage
 import { Playground } from 'docz';
 import { useState } from 'react';
 
-import ReactWordcloud from '../ReactWordcloud';
+import ReactWordcloud from '../react-wordcloud';
 import words from '../words';
 
 # Transitions
